Expose undo/redo hotkeys globally
I don't see a reason why they should only be exposed in the outline panel.
