+++ apps/lite/ui/src/hotkeys.ts
@@ -73,9 +73,17 @@ export const globalHotkeys = {
 		hotkey: "Mod+K",
 		meta: { group: "Global", name: "Command palette" },
 	},
+	redo: {
+		hotkey: "Mod+Shift+Z",
+		meta: { group: "Outline", name: "Redo" },
+	},
 	selectProject: {
 		hotkey: "Mod+Shift+P",
 		meta: { group: "Global", name: "Select project" },
+	},
+	undo: {
+		hotkey: "Mod+Z",
+		meta: { group: "Outline", name: "Undo" },
 	},
 } satisfies Record<string, HotkeyWithMeta>;
 
@@ -127,10 +135,6 @@ export const outlineHotkeys = {
 		hotkey: "Alt+ArrowUp",
 		meta: { group: "Commit", name: "Move commit up" },
 	},
-	redo: {
-		hotkey: "Mod+Shift+Z",
-		meta: { group: "Outline", name: "Redo" },
-	},
 	renameBranch: {
 		hotkey: "Enter",
 		meta: { group: "Branch", name: "Rename branch" },
@@ -146,10 +150,6 @@ export const outlineHotkeys = {
 	selectChanges: {
 		hotkey: "Z",
 		meta: { group: "Outline", name: "Select changes" },
-	},
-	undo: {
-		hotkey: "Mod+Z",
-		meta: { group: "Outline", name: "Undo" },
 	},
 } satisfies Record<string, HotkeyWithMeta>;
